refactor: modify domain package name
- change domain to model.
