+++ src/main/kotlin/model/Computer.kt
@@ -1,3 +1,3 @@
-package domain
+package model
 
 class Computer(numbers: List<Int>) : Role(numbers)
