+++ src/main/kotlin/model/Player.kt
@@ -1,3 +1,3 @@
-package domain
+package model
 
 class Player(numbers: List<Int>) : Role(numbers)
