+++ src/main/kotlin/model/Role.kt
@@ -1,4 +1,4 @@
-package domain
+package model
 
 import utils.Constants
 import utils.ExceptionHandler.validateDuplicatedOrException
