+++ src/test/kotlin/baseball/hint/BaseballHintTest.kt
@@ -1,6 +1,6 @@
 package baseball.hint
 
-import domain.Computer
+import model.Computer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
